Divided views file into parts
