+++ blog/urls.py
@@ -1,35 +1,49 @@
 from django.urls import path
-from . import views
+from .views import homepage_view, content_view, user_view, verification_view
 
 
 app_name = "blog"
 urlpatterns = [
-    path("", views.homepage, name="homepage"),
-    path("create-post/", views.create_post, name="create_post"),
-    path("category/", views.categories_list, name="categories_list"),
-    path("category/<str:cat>/", views.category_details, name="category_details"),
-    path("register/", views.register_user, name="register_user"),
-    path("verify/<uidb64>/<token>", views.verify_email, name="verify_email"),
-    path("login/", views.login_user, name="login_user"),
-    path("recover", views.recover_password_request, name="recover_password"),
+    # HOMEPAGE URLS
+    path("", homepage_view.homepage, name="homepage"),
+    path("category/", homepage_view.categories_list, name="categories_list"),
     path(
-        "recover/<uidb64>/<token>",
-        views.recover_password_confirm,
-        name="recover_password_confirm",
+        "category/<str:cat>/", homepage_view.category_details, name="category_details"
     ),
-    path("logout/", views.logout_user, name="logout_user"),
-    path("dashboard/", views.profile_dashboard, name="profile_dashboard"),
-    path("settings/", views.settings_user, name="settings_user"),
-    path("profile-settings/", views.profile_settings_user, name="profile_settings"),
+    #
+    #
+    # CONTENT URLS
+    path("<slug:slug>/", content_view.post_details, name="post_details"),
+    path("create-post/", content_view.create_post, name="create_post"),
+    path("<slug:slug>/edit-post/", content_view.edit_post, name="edit_post"),
+    path("<slug:slug>/delete-post/", content_view.delete_post, name="delete_post"),
+    path("<int:id>/edit-comment", content_view.edit_comment, name="edit_comment"),
+    path("<int:id>/delete-comment", content_view.delete_comment, name="delete_comment"),
+    #
+    #
+    # VERIFICATION URLS
+    path("register/", verification_view.register_user, name="register_user"),
+    path(
+        "verify/<uidb64>/<token>", verification_view.verify_email, name="verify_email"
+    ),
+    path("login/", verification_view.login_user, name="login_user"),
+    path("logout/", verification_view.logout_user, name="logout_user"),
+    #
+    #
+    # USER URLS
+    path("settings/", user_view.settings_user, name="settings_user"),
     path(
         "notifications/",
-        views.notifications,
+        user_view.notifications,
         name="notifications",
     ),
-    path("change-password/", views.change_password, name="change_password"),
-    path("<int:id>/edit-comment", views.edit_comment, name="edit_comment"),
-    path("<int:id>/delete-comment", views.delete_comment, name="delete_comment"),
-    path("<slug:slug>/", views.post_details, name="post_details"),
-    path("<slug:slug>/edit-post/", views.edit_post, name="edit_post"),
-    path("<slug:slug>/delete-post/", views.delete_post, name="delete_post"),
+    path("profile-settings/", user_view.profile_settings_user, name="profile_settings"),
+    path("dashboard/", user_view.profile_dashboard, name="profile_dashboard"),
+    path("change-password/", user_view.change_password, name="change_password"),
+    path("recover", user_view.recover_password_request, name="recover_password"),
+    path(
+        "recover/<uidb64>/<token>",
+        user_view.recover_password_confirm,
+        name="recover_password_confirm",
+    ),
 ]
